requirements.txt
anyascii==0.3.2
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.13.3
binaryornot==0.4.4
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
cookiecutter==2.6.0
cryptography==44.0.0
defusedxml==0.7.1
Django==5.1.4
django-allauth==65.3.0
django-filter==25.1
django-htmx==1.21.0
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.1.3
django-taggit==6.1.0
django-tailwind==3.8.0
django-tasks==0.6.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
Jinja2==3.1.5
laces==0.1.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
openpyxl==3.1.5
pillow==11.1.0
pillow_heif==0.21.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stripe==11.4.1
telepath==0.3.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
wagtail==6.4.1
whitenoise==6.8.2
Willow==1.9.0
